backend/requirements.txt
pymysql
llama_index
google.generativeai
dotenv
sklearn
numpy
redis